% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendmail.R
\name{sendmail}
\alias{sendmail}
\title{sendmail}
\usage{
sendmail(from, to, subject, msg, smtpsettings, attachment = NULL)
}
\arguments{
\item{from}{From address.}

\item{to}{To address}

\item{subject}{subject of email.}

\item{msg}{character vector holding lines of email or character string with entire message.}

\item{smtpsettings}{server settings for the smtp server. A list with the following entries:
\describe{
  \item{server:}{address of smtp server, e.g. smtp.googlemail.com}
  \item{port:}{optional port, defaults to 465 (SSL)}
  \item{username:}{If server requires authentication, supply username}
  \item{password:}{password if required}
  \item{usessl:}{If port is set to be different from 465, ssl will not be used by default, can be specified here optionally.}
}}

\item{attachment}{optional character vector containing path of files to be attached.}
}
\description{
Send e-mails from R via SMTP with authentication and encryption
}
